fix: propagate executor errors directly to the caller instead of emitting synthetic failure tasks
